requirements.txt
Django==4.2
djangorestframework==3.14.0
Channels==4.0.0
Channels-redis==4.1.0
Daphne==4.0.0
python-dotenv==1.0.1
asyncpg==0.29.0
aioimaplib==1.1.0
pytz==2024.1